Allow specifying additional ninja cflags through an external SKIA_EXTRA_CFLAGS environment variable
This enables the use of clang when it fails to detect the correct gcc
toolchain when cross-compiling. After all, clang uses heuristics, which
may fail. So in order to make the gcc toolchain selection deterministic,
propagate SKIA_EXTRA_CFLAGS to extra_cflags in the ninja invocation, in
order to be able to pass --gcc-toolchain=/path/to/toolchain to clang.